Move tls secrets handling into controller
